+++ Demo/Tests/TestPlugin.swift
@@ -6,25 +6,25 @@ final class TestingPlugin: PluginType {
     var result: Result<Moya.Response, Moya.Error>?
     var didPrepare = false
 
-    func prepareRequest(_ request: URLRequest, target: TargetType) -> URLRequest {
+    func prepare(_ request: URLRequest, target: TargetType) -> URLRequest {
         var request = request
         request.addValue("yes", forHTTPHeaderField: "prepared")
         return request
     }
 
-    func willSendRequest(_ request: RequestType, target: TargetType) {
+    func willSend(_ request: RequestType, target: TargetType) {
         self.request = (request, target)
 
-        // We check for whether or not we did prepare here to make sure prepareRequest gets called
-        // before willSendRequest
+        // We check for whether or not we did prepare here to make sure prepare gets called
+        // before willSend
         didPrepare = request.request?.allHTTPHeaderFields?["prepared"] == "yes"
     }
 
-    func didReceiveResponse(_ result: Result<Moya.Response, Moya.Error>, target: TargetType) {
+    func didReceive(_ result: Result<Moya.Response, Moya.Error>, target: TargetType) {
         self.result = result
     }
 
-    func processResponse(_ result: Result<Response, Moya.Error>, target: TargetType) -> Result<Response, Moya.Error> {
+    func process(_ result: Result<Response, Moya.Error>, target: TargetType) -> Result<Response, Moya.Error> {
         var result = result
 
         if case .success(let response) = result {